Checks for weather and day/night changes on loop
The website now checks for weather and day/night changes whenever the video ends and is about to be looped. Now, a user can simply go to the end of the video if they want the song version to be updated more quickly.
I moved around some of the functions so that similar ones can be found together. I also renamed some of the functions to better fit their behavior, and added one, refreshPlayer(), to allow for different cases where the video has to be refreshed (when the video loops, and when new weather conditions are found after 5 minutes.)