Initial changes for the v2.0 update
A few important changes.

One big change is that, for the time being, the website will utilize the Tomorrow.io weather API. Although the OpenWeatherMap API worked, it wasn't always the most accurate, as I believe the free current weather data only updates every 2 hours. The Tomorrow.io API is much more accurate. Although I can make way less free calls, the 500 per day is enough for my own personal use. The question of what API to use will definitely come back if I ever host this website publicly, though.

Another important change is that the website no longer asks for the user's location as soon as they enter the site. Alternatively, the user can enable weather-tracking if they click on the sun icon. This change was done in an effort to get the website closer to a "production" version, i.e. something that could be acceptably published/hosted on the Internet taking into account user concerns about privacy and security. A user would probably get spooked if they saw a website immediately asking for their location.

I also added in support for the Arsenal font, though it's not in use anywhere on the site yet. It's supposed to be pretty similar to the font used for the dialogue in Persona 5.