Added proof of concept for calendar element
For now, text will display in the top-left corner that displays the current date and day of the week. I created a new script file and edited the HTML a bit as well for this feature. The plan is to develop a system that will recreate the calendar HUD element in Persona 5.