An EXTREMELY rough attempt at rendering calendar
I'm able to render the month and date using a canvas (not the day of the week yet), but the actual proportions and orientations of the numbers and backgrounds are off. Also, the way I'm loading in the images is not the most efficient. Nonetheless, I've shown that a calendar element is feasible. I still need to implement functionality for having the calendar graphic change when the date changes.